nrf52_bluetooth: split PPoGATT into its own file
